Move the api/grades/earned_badges routes to a sub-controller
